+++ packages/common/src/index.ts
@@ -9,7 +9,7 @@ import * as Constants from "./constants";
 import * as Types from "./types";
 import * as Util from "./util";
 import * as Constraints from "./constraints/constraints";
-import * as Score from "./score";
+import * as Score from "./constraints/score";
 import * as Scripts from "./scripts";
 import * as Models from "./db/models";
